+++ ios_app/Features/schoolCard/Extensions/Extentions.swift
